Fix MSVC warnings about integer size changes and data loss

+++ src/corelib/tools/qarraydata.cpp
@@ -138,7 +138,7 @@ void *QArrayData::allocate(QArrayData **dptr, size_t objectSize, size_t alignmen
         // find where offset should point to so that data() is aligned to alignment bytes
         data = (quintptr(header) + sizeof(QArrayAllocatedData) + alignment - 1)
                 & ~(alignment - 1);
-        header->alloc = capacity;
+        header->alloc = uint(capacity);
     }
 
     *dptr = header;
@@ -165,7 +165,7 @@ QArrayData::reallocateUnaligned(QArrayData *data, void *dataPointer,
     options |= AllocatedDataType | MutableData;
     QArrayAllocatedData *header = static_cast<QArrayAllocatedData *>(reallocateData(data, allocSize, options));
     if (header) {
-        header->alloc = capacity;
+        header->alloc = uint(capacity);
         dataPointer = reinterpret_cast<char *>(header) + offset;
     }
     return qMakePair(static_cast<QArrayData *>(header), dataPointer);
